Add basic support for external unit tests
Add new test type of "external" (name may be changed in future), which
is largely inspired by Gradescope. An arbitrary command can be executed,
which should produce a JSON file summarizing autograde test results.

Currently, only the top-level "score" and "max_score" JSON fields are
supported. If *both* of those are present, then they are used to update
the total and possible points ouput by the autograding action
(otherwise, it is considered an ungraded test).